Refactor SettingsRepository into traits
